issue: 1048224 Use ioctl() to verify if non-offloaded data is available
Insted of using os.poll() to verify if non-offloaded data is available,
We can use os.ioctl() using FIONREAD flag which takes less time.

+++ src/vma/sock/sockinfo_udp.h
@@ -265,7 +265,7 @@ private:
 	int 		rx_wait_helper(int &poll_count, bool is_blocking);
 	
 	inline int 	rx_wait(bool blocking);
-	inline ssize_t	poll_os();
+	inline int 	poll_os();
 
 	virtual inline void			reuse_buffer(mem_buf_desc_t *buff);
 	virtual 	mem_buf_desc_t*	get_next_desc (mem_buf_desc_t *p_desc);
